feat: support use custom scheduler plugins and do some refact
